add policies and fix validation

+++ app/Http/Controllers/TaskController.php
@@ -31,6 +31,7 @@ class TaskController extends Controller
      */
     public function store(TaskRequest $request)
     {
+        $this->authorize('create', Task::class);
         return Task::create($request->all());
     }
 
@@ -42,6 +43,7 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        $this->authorize('view', $task);
         return view('show', compact(['task']));
     }
 
@@ -54,6 +56,7 @@ class TaskController extends Controller
      */
     public function update(TaskRequest $request, Task $task)
     {
+        $this->authorize('update', $task);
         if ($request->done) {
             $task->update([
                 'done' => ($task->done == 0) ? 1 : 0,
@@ -70,6 +73,7 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        $this->authorize('forceDelete', $task);
         if ($task->delete()) {
             return response(['message' => 'deleted'], 200);
         }
